+Updated the rotatable platform so that the player can attach to it and spin with it, the player is also able to dash out of it.

+++ Assets/Scripts/MiscScripts/LogoTimer.cs
@@ -5,7 +5,7 @@ using UnityEngine.SceneManagement;
 
 public class LogoTimer : MonoBehaviour {
 
-	public int timer;
+	public float timer;
 
 	void Start ()
 	{
@@ -16,4 +16,5 @@ public class LogoTimer : MonoBehaviour {
 	{
 		SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex + 1);
 	}
+
 }
